bugfix: исправление ошибок в автотестах #3

+++ controllers/cards.js
@@ -1,6 +1,7 @@
 const Card = require('../models/Card');
 const { BadRequestError } = require('../errors/bad-request-err');
 const { NotFoundError } = require('../errors/not-found-err');
+const { ForbiddenError } = require('../errors/forbidden-err');
 const { ServerError } = require('../errors/server-err');
 
 const getCards = async (req, res, next) => {
@@ -20,7 +21,7 @@ const createCard = async (req, res, next) => {
     if (e.name === 'ValidationError') {
       return next(new BadRequestError('Ошибка в запросе'));
     }
-    next();
+    return next();
   }
 };
 
@@ -28,16 +29,22 @@ const deleteCardById = async (req, res, next) => {
   const { cardId } = req.params;
   const currentUserId = req.user._id;
   try {
-    const card = await Card.findByIdAndDelete(cardId);
-    if (!card) {
+    if (!cardId) {
       return next(new NotFoundError('карточка не найдена'));
     }
+    const card = await Card.findById(cardId);
+    const cardOwner = await Card.findByIdAndDelete(cardId);
+    if (cardOwner !== currentUserId) {
+      return next(new ForbiddenError('Не хватает прав на удаление чужой карточки!'));
+    }
+    await Card.findByIdAndDelete(cardId);
+
     return res.status(200).send(card);
   } catch (e) {
     if (e.name === 'CastError') {
       return next(new BadRequestError('Ошибка в запросе'));
     }
-    next();
+    return next();
   }
 };
 
@@ -56,7 +63,7 @@ const likeCard = async (req, res, next) => {
     if (e.name === 'CastError') {
       return next(new BadRequestError('Ошибка в запросе'));
     }
-    next();
+    return next();
   }
 };
 
@@ -75,7 +82,7 @@ const dislikeCard = async (req, res, next) => {
     if (e.name === 'CastError') {
       return next(new BadRequestError('Ошибка в запросе'));
     }
-    next();
+    return next();
   }
 };
